backend/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
distro==1.8.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
peewee==3.16.2
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tabula-py==2.7.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
